rename: 메인 액티비티 이름 수정
- MainActivity.kt -> MovieActivity.kt
- MainActivityTest.kt -> MovieActivityTest.kt
