+++ app/src/main/res/layout/activity_movie.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.MainActivity">
+    tools:context=".ui.view.movie.MovieActivity">
 
     <ListView
         android:id="@+id/listview_layout"
